feat(images): run sra-parser on stock requests biocontainer via bin/
The sra-parser custom image bundled requests + beautifulsoup4 + lxml + html5lib,
none of which has a stock biocontainer (bioconda does not package beautifulsoup4
or html5lib), so there is no drop-in biocontainer for that stack.

Rather than keep a custom docker.io/eoksen image (which the identity migration
would break), reduce the scraper to its essentials: it now uses `requests` for
the HTTP calls (unchanged) and the Python standard library `xml.etree` for XML
parsing instead of BeautifulSoup/lxml/html5lib. Behaviour is preserved — same
esearch/efetch endpoints, same rate limiting, same `<srr>,<identifier>` CSV
output (verified against live NCBI for SRR18506022 -> SRR18506022,NC_000866.4).

With only `requests` remaining, the script runs under the stock
quay.io/biocontainers/requests image. The script moves to the top-level bin/
(auto-staged onto every task PATH by Nextflow) with a shebang, and get_srrs /
parse_srrs invoke it as `sra_parser.py`. The custom Dockerfile and build block
are removed; the manifest entry is now runtime-only.
